Code review improvements
- Move `aspectRatio` to `SizeConstrainableProxy`
